check inference
