+++ utils.py
@@ -11,13 +11,3 @@ def seed_everything(seed):
     torch.cuda.manual_seed(seed)
     torch.backends.cudnn.deterministic = True
     torch.backends.cudnn.benchmark = True
-def post_processing(predictions):
-    print(predictions.shape) ##(2000000,2)
-    softmax = torch.nn.Softmax(dim=-1)
-    predictions = torch.tensor(predictions)
-    pred_score = softmax(predictions)
-    pred_score = pred_score.view(-1, 10, 2)
-    pred_score = pred_score.numpy()
-    pred_score = pred_score[:,:,1]
-    preds = np.argmax(pred_score, axis=-1)
-    return preds
